Works phase shift (looower freq) - only at startup (not inside while)

+++ HRtimer/Src/utils.h
@@ -1,7 +1,8 @@
 #ifndef __UTILS_H__
 #define __UTILS_H__
 
+#include <stdint.h>
 void advanced_int_init(void);
-void update_phase(int val);
+void update_phase(uint16_t val);
 
 #endif
